Move mocking functions
